separate Conv2dSubsampling from Zipformer
