+++ egs/librispeech/ASR/pruned_transducer_stateless7/decode.py
@@ -127,6 +127,7 @@ from icefall.checkpoint import (
 from icefall.lexicon import Lexicon
 from icefall.utils import (
     AttributeDict,
+    make_pad_mask,
     setup_logger,
     store_transcripts,
     str2bool,
@@ -365,9 +366,15 @@ def decode_one_batch(
             value=LOG_EPS,
         )
 
+    x, x_lens = model.encoder_embed(feature, feature_lens)
+
+    src_key_padding_mask = make_pad_mask(x_lens)
+    x = x.permute(1, 0, 2)  # (N, T, C) -> (T, N, C)
+
     encoder_out, encoder_out_lens = model.encoder(
-        x=feature, x_lens=feature_lens
+        x, x_lens, src_key_padding_mask
     )
+    encoder_out = encoder_out.permute(1, 0, 2)  # (T, N, C) ->(N, T, C)
 
     hyps = []
